+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: sadawi <sadawi@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/18 14:54:37 by sadawi            #+#    #+#             */
-/*   Updated: 2019/10/28 18:57:41 by sadawi           ###   ########.fr       */
+/*   Updated: 2019/11/06 13:47:00 by sadawi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,13 @@
 
 void	*ft_memmove(void *dst, const void *src, size_t len)
 {
-	int		i;
+	size_t	i;
 
 	if (!dst && !src)
 		return (dst);
 	i = 0;
 	if (src > dst)
-		while (i < (int)len)
+		while (i < len)
 		{
 			((unsigned char*)dst)[i] = ((unsigned char*)src)[i];
 			i++;
@@ -28,7 +28,7 @@ void	*ft_memmove(void *dst, const void *src, size_t len)
 	else
 	{
 		i = 1;
-		while (i <= (int)len)
+		while (i <= len)
 		{
 			((unsigned char*)dst)[len - i] = ((unsigned char*)src)[len - i];
 			i++;
