+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: sadawi <sadawi@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/18 12:23:13 by sadawi            #+#    #+#             */
-/*   Updated: 2019/10/18 14:02:10 by sadawi           ###   ########.fr       */
+/*   Updated: 2019/11/06 13:47:25 by sadawi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 void	*ft_memset(void *b, int c, size_t len)
 {
-	int i;
+	size_t i;
 
 	i = 0;
-	while (i < (int)len)
+	while (i < len)
 	{
 		((unsigned char *)b)[i] = (unsigned char)c;
 		i++;
