+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: sadawi <sadawi@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/17 17:53:01 by sadawi            #+#    #+#             */
-/*   Updated: 2019/10/23 17:08:59 by sadawi           ###   ########.fr       */
+/*   Updated: 2019/11/06 13:54:04 by sadawi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int i;
-	int j;
+	size_t i;
+	size_t j;
 
 	i = 0;
 	j = 0;
@@ -23,7 +23,7 @@ char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 		return ((char*)haystack);
 	while (haystack[i])
 	{
-		while (haystack[i + j] == needle[j] && i + j < (int)len)
+		while (haystack[i + j] == needle[j] && i + j < len)
 		{
 			if (!needle[j + 1])
 				return ((char*)&haystack[i]);
