+++ ft_strsub.c
@@ -6,7 +6,7 @@
 /*   By: sadawi <sadawi@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/22 15:20:24 by sadawi            #+#    #+#             */
-/*   Updated: 2019/10/25 14:23:48 by sadawi           ###   ########.fr       */
+/*   Updated: 2019/11/06 13:55:21 by sadawi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 char	*ft_strsub(char const *s, unsigned int start, size_t len)
 {
 	char	*str;
-	int		i;
+	size_t	i;
 
 	i = 0;
 	if (!(str = (char*)malloc(len + 1)))
